requirements.txt
asgiref==3.7.2
boto3==1.28.78
botocore==1.31.78
certifi==2023.7.22
charset-normalizer==3.3.0
dj-database-url==2.2.0
Django==4.2.5
django-cors-headers==4.2.0
django-crontab==0.7.1
django-heroku==0.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==22.0.0
idna==3.4
jmespath==1.0.1
joblib==1.3.2
numpy==1.26.1
packaging==24.1
psycopg2==2.9.9
PyJWT==2.8.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
urllib3==1.25.4
whitenoise==6.7.0
